# Optimize recomputing definition