--- a/lib/utils.js
+++ b/lib/utils.js
@@ -88,6 +88,7 @@ function getLatestVersion(versions) {
     const normalizedCurrent = _normalizeVersion(current)
     if (!normalizedCurrent || semver.prerelease(normalizedCurrent) !== null) return max
     const normalizedMax = _normalizeVersion(max)
+    if (!normalizedMax) return normalizedCurrent
     return semver.gt(normalizedCurrent, normalizedMax) ? current : max
   }, versions[0])
 }