… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/clearlydefined/1.1.2.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/clearlydefined/1.3.4.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/clearlydefined/1.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/fossology/3.3.0.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/fossology/3.6.0.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/licensee/9.12.1.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/licensee/9.14.0.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/reuse/1.3.0.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/reuse/3.2.1.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/scancode/2.2.1.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/scancode/2.9.0+b1.json
… test/fixtures/store/npm/npmjs/-/debug/revision/3.1.0/tool/scancode/30.3.0.json